# api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoCertX Control Plane and Agent API",
    "version": "1.0.1-t03",
    "description": "Initial frozen API contract for GA phase one. Covers control plane REST APIs, agent REST protocol, and shared error semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": ["jobs"],
        "operationId": "listJobs",
        "summary": "List jobs",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Job list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "tags": ["jobs"],
        "operationId": "getJob",
        "summary": "Get job detail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Job detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/jobs/{id}/attempts": {
      "get": {
        "tags": ["jobs"],
        "operationId": "listJobAttempts",
        "summary": "List job attempts",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Job attempt list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/api/v1/dashboard/summary": {
      "get": {
        "tags": ["dashboard"],
        "operationId": "getDashboardSummary",
        "summary": "Get dashboard summary",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/api/v1/dashboard/job-failures": {
      "get": {
        "tags": ["dashboard"],
        "operationId": "listDashboardJobFailures",
        "summary": "List dashboard job failures",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Job failure list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": { "type": "string" },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 }
        }
      },
      "ObjectEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": { "type": "string" },
          "data": { "type": "object" }
        }
      },
      "ListEnvelope": {
        "type": "object",
        "required": ["request_id", "items"],
        "properties": {
          "request_id": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "type": "object" }
          },
          "meta": { "$ref": "#/components/schemas/PageMeta" }
        }
      },
...
    }
  }
}